add functions to delete site user and site group

+++ src/get_data.py
@@ -8,9 +8,10 @@
 import logging
 import json
 import re
+import uuid
 from datetime import datetime, timezone
 from typing import List, Optional
-
+import sqlalchemy as sa
 from sqlalchemy.orm.session import Session
 from sqlalchemy.sql.functions import func
 from sqlalchemy.orm import joinedload
@@ -22,7 +23,7 @@ from nowcasting_datamodel.models.metric import (
     MetricSQL,
     MetricValueSQL,
 )
-from pvsite_datamodel.sqlmodels import UserSQL, SiteGroupSQL, SiteSQL
+from pvsite_datamodel.sqlmodels import UserSQL, SiteGroupSQL, SiteSQL, ForecastValueSQL, ForecastSQL
 
 from data.gsp import get_gsp
 from data.dno import get_dno
@@ -310,6 +311,7 @@ def create_user(
 
     return user
 
+
 # delete a site 
 def delete_site(session: Session, site_uuid: str) -> SiteGroupSQL:
     """Delete a site group.
@@ -317,6 +319,19 @@ def delete_site(session: Session, site_uuid: str) -> SiteGroupSQL:
     :param site_uuid: unique identifier for site
     """
     site = session.query(SiteSQL).filter(SiteSQL.site_uuid == site_uuid).first()
+
+    forecast_uuids = session.query(ForecastSQL.forecast_uuid).filter(ForecastSQL.site_uuid == site_uuid).all()
+
+    forecast_uuids = [str(forecast_uuid[0]) for forecast_uuid in forecast_uuids]
+
+    # get and delete all forecast values for site 
+    stmt = sa.delete(ForecastValueSQL).where(ForecastValueSQL.forecast_uuid.in_(forecast_uuids))
+    session.execute(stmt)
+
+    stmt_2 = sa.delete(ForecastSQL).where(ForecastSQL.forecast_uuid.in_(forecast_uuids))
+    session.execute(stmt_2)
+
+    # we decided not to delete generation data for the site because it seems sensible to keep it for now
 
     session.delete(site)
 
@@ -336,7 +351,7 @@ def delete_user(session: Session, email: str) -> UserSQL:
 
     session.delete(user)
     
-    message = f"User with email {user.email} and user uuid {user.user_uuid} deleted successfully"
+    message = f"User with email {user.email} and site_group_uuid {user.site_group_uuid} deleted successfully"
 
     session.commit()
 
@@ -354,9 +369,15 @@ def delete_site_group(session: Session, site_group_name: str) -> SiteGroupSQL:
         .first()
     )
 
+    site_group_users = site_group.users
+
+    if len(site_group_users) > 0:
+        message = f"Site group with name {site_group.site_group_name} and site group uuid {site_group.site_group_uuid} has users and cannot be deleted."
+        return message
+    
     session.delete(site_group)
 
-    message = f"Site group with name {site_group.site_group_name} and site group uuid {site_group.site_group_uuid} deleted successfully"
+    message = f"Site group with name {site_group.site_group_name} and site group uuid {site_group.site_group_uuid} deleted successfully."
 
     session.commit()
